marryme-bot/requirements.txt
Flask>=2.0
Pillow>=9.0
google-cloud-storage>=2.0
gunicorn>=20.0
line-bot-sdk>=3.0.0
pypinyin==0.54.0